Remove optimized naming from contracts and tests
