+++ contracts/TESTForceMove.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.5.11;
 pragma experimental ABIEncoderV2;
-import './OptimizedForceMove.sol';
+import './ForceMove.sol';
 
-contract TESTOptimizedForceMove is OptimizedForceMove {
+contract TESTForceMove is ForceMove {
     // Public wrappers for internal methods:
 
     function isAddressInArray(address suspect, address[] memory addresses)
