+++ test/forceMove/_validSignatures.ts
@@ -1,12 +1,12 @@
 import {ethers} from 'ethers';
 // @ts-ignore
-import OptimizedForceMoveArtifact from '../../build/contracts/TESTOptimizedForceMove.json';
+import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
 import {setupContracts, sign} from './test-helpers';
 
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
 );
-let OptimizedForceMove: ethers.Contract;
+let ForceMove: ethers.Contract;
 
 const participants = ['', '', ''];
 const wallets = new Array(3);
@@ -18,7 +18,7 @@ for (let i = 0; i < 3; i++) {
 }
 
 beforeAll(async () => {
-  OptimizedForceMove = await setupContracts(provider, OptimizedForceMoveArtifact);
+  ForceMove = await setupContracts(provider, ForceMoveArtifact);
 });
 
 // TODO use .each to improve readability and reduce boilerplate
@@ -47,18 +47,12 @@ describe('_validSignatures', () => {
       sigs[i] = {v: sig.v, r: sig.r, s: sig.s};
       whoSignedWhat[i] = i;
     }
-    expect(
-      await OptimizedForceMove.validSignatures(8, addresses, stateHashes, sigs, whoSignedWhat),
-    ).toBe(true);
+    expect(await ForceMove.validSignatures(8, addresses, stateHashes, sigs, whoSignedWhat)).toBe(
+      true,
+    );
     const brokenSigs = sigs.reverse();
     expect(
-      await OptimizedForceMove.validSignatures(
-        8,
-        addresses,
-        stateHashes,
-        brokenSigs,
-        whoSignedWhat,
-      ),
+      await ForceMove.validSignatures(8, addresses, stateHashes, brokenSigs, whoSignedWhat),
     ).toBe(false);
   });
   it('returns true (false) for a correct (incorrect) set of signatures on 1 state', async () => {
@@ -70,18 +64,12 @@ describe('_validSignatures', () => {
       sigs[i] = {v: sig.v, r: sig.r, s: sig.s};
       whoSignedWhat[i] = 0;
     }
-    expect(
-      await OptimizedForceMove.validSignatures(8, addresses, [stateHash], sigs, whoSignedWhat),
-    ).toBe(true);
+    expect(await ForceMove.validSignatures(8, addresses, [stateHash], sigs, whoSignedWhat)).toBe(
+      true,
+    );
     const brokenSigs = sigs.reverse();
     expect(
-      await OptimizedForceMove.validSignatures(
-        8,
-        addresses,
-        [stateHash],
-        brokenSigs,
-        whoSignedWhat,
-      ),
+      await ForceMove.validSignatures(8, addresses, [stateHash], brokenSigs, whoSignedWhat),
     ).toBe(false);
   });
 });
