+++ test/forceMove/respond.test.ts
@@ -1,7 +1,7 @@
 import {ethers} from 'ethers';
 import {expectRevert} from 'magmo-devtools';
 // @ts-ignore
-import OptimizedForceMoveArtifact from '../../build/contracts/TESTOptimizedForceMove.json';
+import ForceMoveArtifact from '../../build/contracts/TESTForceMove.json';
 // @ts-ignore
 import countingAppArtifact from '../../build/contracts/CountingApp.json';
 import {keccak256, defaultAbiCoder, hexlify} from 'ethers/utils';
@@ -11,7 +11,7 @@ import {HashZero, AddressZero} from 'ethers/constants';
 const provider = new ethers.providers.JsonRpcProvider(
   `http://localhost:${process.env.DEV_GANACHE_PORT}`,
 );
-let OptimizedForceMove: ethers.Contract;
+let ForceMove: ethers.Contract;
 let networkId;
 const chainId = 1234;
 const participants = ['', '', ''];
@@ -29,41 +29,39 @@ for (let i = 0; i < 3; i++) {
 const nonParticipant = ethers.Wallet.createRandom();
 
 beforeAll(async () => {
-  OptimizedForceMove = await setupContracts(provider, OptimizedForceMoveArtifact);
+  ForceMove = await setupContracts(provider, ForceMoveArtifact);
   networkId = (await provider.getNetwork()).chainId;
   appDefinition = countingAppArtifact.networks[networkId].address; // use a fixed appDefinition in all tests
 });
 
 // Scenarios are synonymous with channelNonce:
 
-const description1 = 'It accepts a refute tx for an ongoing challenge';
-const description2 = 'It reverts a refute tx if the challenge has expired';
-const description3 = 'It reverts a refute tx if the declaredTurnNumRecord is incorrect';
-const description4 =
-  'It reverts a refute tx if the refutation state is not signed by the challenger';
+const description1 = 'It accepts a respond tx for an ongoing challenge';
+const description2 = 'It reverts a respond tx if the challenge has expired';
+const description3 = 'It reverts a respond tx if the declaredTurnNumRecord is incorrect';
+const description4 = 'It reverts a respond tx if it is not signed by the correct participant';
 const description5 =
-  'It reverts a refute tx if the refutationTurnNum is not larger than declaredTurnNumRecord';
+  'It reverts a respond tx if the response state is not a validTransition from the challenge state';
 
 describe('respond', () => {
   it.each`
-    description     | channelNonce | setTurnNumRecord | declaredTurnNumRecord | refutationTurnNum | expired  | isFinalAB         | appDatas  | challenger    | refutationStateSigner | reasonString
-    ${description1} | ${1001}      | ${8}             | ${8}                  | ${14}             | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[2]}         | ${undefined}
-    ${description2} | ${1002}      | ${8}             | ${8}                  | ${14}             | ${true}  | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[2]}         | ${'Refute too late!'}
-    ${description3} | ${1003}      | ${8}             | ${7}                  | ${14}             | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[2]}         | ${'Challenge State does not match stored version'}
-    ${description4} | ${1004}      | ${8}             | ${8}                  | ${14}             | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${nonParticipant}     | ${'Refutation state not signed by challenger'}
-    ${description5} | ${1001}      | ${8}             | ${8}                  | ${5}              | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[2]}         | ${'Refutation state must have a higher turn number'}
+    description     | channelNonce | setTurnNumRecord | declaredTurnNumRecord | expired  | isFinalAB         | appDatas  | challenger    | responder         | reasonString
+    ${description1} | ${1001}      | ${8}             | ${8}                  | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[0]}     | ${undefined}
+    ${description2} | ${1002}      | ${8}             | ${8}                  | ${true}  | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[0]}     | ${'Response too late!'}
+    ${description3} | ${1003}      | ${8}             | ${7}                  | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${wallets[0]}     | ${'Challenge State does not match stored version'}
+    ${description4} | ${1004}      | ${8}             | ${8}                  | ${false} | ${[false, false]} | ${[0, 1]} | ${wallets[2]} | ${nonParticipant} | ${'Response not signed by authorized mover'}
+    ${description5} | ${1005}      | ${8}             | ${8}                  | ${false} | ${[false, false]} | ${[0, 0]} | ${wallets[2]} | ${wallets[0]}     | ${'CountingApp: Counter must be incremented'}
   `(
     '$description', // for the purposes of this test, chainId and participants are fixed, making channelId 1-1 with channelNonce
     async ({
       channelNonce,
       setTurnNumRecord,
       declaredTurnNumRecord,
-      refutationTurnNum,
       expired,
       isFinalAB,
       appDatas,
       challenger,
-      refutationStateSigner,
+      responder,
       reasonString,
     }) => {
       // compute channelId
@@ -106,27 +104,27 @@ describe('respond', () => {
         ),
       );
 
-      const refutationAppPartHash = keccak256(
+      const responseAppPartHash = keccak256(
         defaultAbiCoder.encode(
           ['uint256', 'address', 'bytes'],
           [challengeDuration, appDefinition, defaultAbiCoder.encode(['uint256'], [appDatas[1]])],
         ),
       );
 
-      const refutationState = {
-        turnNum: refutationTurnNum,
+      const responseState = {
+        turnNum: setTurnNumRecord + 1,
         isFinal: isFinalAB[1],
         channelId,
-        refutationAppPartHash,
+        responseAppPartHash,
         outcomeHash,
       };
 
-      const refutationStateHash = keccak256(
+      const responseStateHash = keccak256(
         defaultAbiCoder.encode(
           [
-            'tuple(uint256 turnNum, bool isFinal, bytes32 channelId, bytes32 refutationAppPartHash, bytes32 outcomeHash)',
+            'tuple(uint256 turnNum, bool isFinal, bytes32 channelId, bytes32 responseAppPartHash, bytes32 outcomeHash)',
           ],
-          [refutationState],
+          [responseState],
         ),
       );
 
@@ -134,7 +132,7 @@ describe('respond', () => {
         outcome,
         appData: defaultAbiCoder.encode(['uint256'], [appDatas[0]]), // a counter
       };
-      const refutationVariablePart = {
+      const responseVariablePart = {
         outcome,
         appData: defaultAbiCoder.encode(['uint256'], [appDatas[1]]), // a counter
       };
@@ -155,42 +153,42 @@ describe('respond', () => {
       );
 
       // call public wrapper to set state (only works on test contract)
-      const tx = await OptimizedForceMove.setChannelStorageHash(channelId, challengeExistsHash);
+      const tx = await ForceMove.setChannelStorageHash(channelId, challengeExistsHash);
       await tx.wait();
-      expect(await OptimizedForceMove.channelStorageHashes(channelId)).toEqual(challengeExistsHash);
+      expect(await ForceMove.channelStorageHashes(channelId)).toEqual(challengeExistsHash);
 
       // sign the state
-      const signature = await sign(refutationStateSigner, refutationStateHash);
-      const refutationStateSig = {v: signature.v, r: signature.r, s: signature.s};
-
-      const challengeClearedEvent: any = newChallengeClearedEvent(OptimizedForceMove, channelId);
+      const signature = await sign(responder, responseStateHash);
+      const sig = {v: signature.v, r: signature.r, s: signature.s};
 
       if (reasonString) {
-        expectRevert(
+        const regex = new RegExp(
+          '^' + 'VM Exception while processing transaction: revert ' + reasonString + '$',
+        );
+        await expectRevert(
           () =>
-            OptimizedForceMove.refute(
+            ForceMove.respond(
               declaredTurnNumRecord,
-              refutationTurnNum,
               finalizesAt,
               challenger.address,
               isFinalAB,
               fixedPart,
-              [challengeVariablePart, refutationVariablePart],
-              refutationStateSig,
+              [challengeVariablePart, responseVariablePart],
+              sig,
             ),
-          'VM Exception while processing transaction: revert ' + reasonString,
+          regex,
         );
       } else {
+        const challengeClearedEvent: any = newChallengeClearedEvent(ForceMove, channelId);
         // call respond
-        const tx2 = await OptimizedForceMove.refute(
+        const tx2 = await ForceMove.respond(
           declaredTurnNumRecord,
-          refutationTurnNum,
           finalizesAt,
           challenger.address,
           isFinalAB,
           fixedPart,
-          [challengeVariablePart, refutationVariablePart],
-          refutationStateSig,
+          [challengeVariablePart, responseVariablePart],
+          sig,
         );
 
         // wait for tx to be mined
@@ -198,19 +196,23 @@ describe('respond', () => {
 
         // catch ChallengeCleared event
         const [_, eventTurnNumRecord] = await challengeClearedEvent;
-        expect(eventTurnNumRecord._hex).toEqual(hexlify(declaredTurnNumRecord));
+        expect(eventTurnNumRecord._hex).toEqual(hexlify(declaredTurnNumRecord + 1));
 
         // compute and check new expected ChannelStorageHash
-        const expectedChannelStorage = [declaredTurnNumRecord, 0, HashZero, AddressZero, HashZero];
+        const expectedChannelStorage = [
+          declaredTurnNumRecord + 1,
+          0,
+          HashZero,
+          AddressZero,
+          HashZero,
+        ];
         const expectedChannelStorageHash = keccak256(
           defaultAbiCoder.encode(
             ['uint256', 'uint256', 'bytes32', 'address', 'bytes32'],
             expectedChannelStorage,
           ),
         );
-        expect(await OptimizedForceMove.channelStorageHashes(channelId)).toEqual(
-          expectedChannelStorageHash,
-        );
+        expect(await ForceMove.channelStorageHashes(channelId)).toEqual(expectedChannelStorageHash);
       }
     },
   );
